rename maths subpackage and DataSet class
